manager/requirements.txt
flask
Flask-SQLAlchemy
flask-login
json
requests
collections
pandas
atplotlib